supu_app/requirements.txt
Pillow==11.2.1
pandas==2.3.0
